change int returns types to int64_t

+++ src/runtime/file.c
@@ -20,7 +20,7 @@ cipr_str_t cipr_fread_all(cipr_str_t path) {
     return (cipr_str_t){ .len = read, .data = buf };
 }
 
-int cipr_fwrite(cipr_str_t path, cipr_str_t content) {
+int64_t cipr_fwrite(cipr_str_t path, cipr_str_t content) {
     FILE *f = fopen(path.data, "w");
     if (!f) return -1;
     size_t written = fwrite(content.data, 1, (size_t)content.len, f);
@@ -28,7 +28,7 @@ int cipr_fwrite(cipr_str_t path, cipr_str_t content) {
     return (written == (size_t)content.len) ? 0 : -1;
 }
 
-int cipr_fappend(cipr_str_t path, cipr_str_t content) {
+int64_t cipr_fappend(cipr_str_t path, cipr_str_t content) {
     FILE *f = fopen(path.data, "a");
     if (!f) return -1;
     size_t written = fwrite(content.data, 1, (size_t)content.len, f);
@@ -36,7 +36,7 @@ int cipr_fappend(cipr_str_t path, cipr_str_t content) {
     return (written == (size_t)content.len) ? 0 : -1;
 }
 
-int cipr_file_exists(cipr_str_t path) {
+int64_t cipr_file_exists(cipr_str_t path) {
     FILE *f = fopen(path.data, "r");
     if (!f) return 0;
     fclose(f);
